+++ Packages/Tracking/CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Android 	v5.17.1
 
 ### Added
-- 
+- LeapServiceProvider accessor for world space psoition of the Tracking Camera
+- LeapXRServiceProvider accessor for world space psoition of the Tracking Camera
 
 ### Changed
 - (Config) Additional uses of Config marked as Obsolete